+++ uk.co.compucorp.civicrm.hrcore/backstop_data/page-objects/tabs/absence/work-patterns.js
@@ -1,0 +1,18 @@
+var tab = require('../tab');
+
+module.exports = tab.extend({
+  readySelector: 'absence-tab-work-patterns table',
+  tabTitle: 'Work Patterns',
+
+  /**
+   * Shows the Custom Work Pattern modal
+   */
+  showModal: function () {
+    var casper = this.casper;
+
+    casper.then(function () {
+      casper.click('[ng-click="workpatterns.openModal()"]');
+      casper.waitUntilVisible('absence-tab-custom-work-pattern-modal .modal-body > .row');
+    });
+  }
+});
