+++ uk.co.compucorp.civicrm.hrcore/backstop_data/scenarios/tab-absence.json
@@ -9,6 +9,21 @@
       "label": "Absence /Calendar / All Months",
       "url": "civicrm/contact/view?reset=1&cid=2&selectedChild=absence",
       "onReadyScript": "contact-summary/absence/tab-calendar-all-months"
+    },
+    {
+      "label": "Absence / Entitlements",
+      "url": "civicrm/contact/view?reset=1&cid=2&selectedChild=absence",
+      "onReadyScript": "contact-summary/absence/tab-entitlements"
+    },
+    {
+      "label": "Absence / Work Patterns",
+      "url": "civicrm/contact/view?reset=1&cid=2&selectedChild=absence",
+      "onReadyScript": "contact-summary/absence/tab-work-patterns"
+    },
+    {
+      "label": "Absence / Work Patterns / Custom Work Patterns Modal",
+      "url": "civicrm/contact/view?reset=1&cid=2&selectedChild=absence",
+      "onReadyScript": "contact-summary/absence/tab-work-patterns-modal"
     }
   ]
 }
